pyproject.toml
[project]
name = "barcode-scanner-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.13.2",
    "anyio==4.6.2.post1",
    "bcrypt==4.2.0",
    "blinker==1.8.2",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "colorama==0.4.6",
    "cryptography==44.0.0",
    "distlib==0.3.9",
    "filelock==3.16.1",
    "flask==3.0.3",
    "flask-bcrypt==1.0.1",
    "flask-cors==5.0.0",
    "flask-jwt-extended==4.6.0",
    "flask-login==0.6.3",
    "flask-migrate==4.0.7",
    "flask-sqlalchemy==3.1.1",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httpx==0.27.2",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.4",
    "mako==1.3.5",
    "markupsafe==2.1.5",
    "numpy==2.1.3",
    "opencv-python-headless==4.11.0.86",
    "packaging==24.2",
    "pandas==2.2.3",
    "platformdirs==4.3.6",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pyjwt==2.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "pyzbar==0.1.9",
    "requests==2.32.3",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.32",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "ujson>=5.10.0",
    "urllib3==2.2.3",
    "virtualenv==20.27.1",
    "werkzeug==3.0.4",
]
